updated failing integration tests, fixed indentation in ppl/DateTimeFunctionIT, renamed file to previous name, added type diagram to docs
