+++ integ-test/src/test/java/org/opensearch/sql/ppl/DateTimeFunctionIT.java
@@ -241,7 +241,7 @@ public class DateTimeFunctionIT extends PPLIntegTestCase {
                   "source=%s | eval f = DATETIME('2008-12-25 05:30:00+00:00', 'America/Los_Angeles') | fields f",
                   TEST_INDEX_DATE));
   verifySchema(result,
-               schema("f", null, "timestamp"));
+        schema("f", null, "timestamp"));
   verifySome(result.getJSONArray("datarows"), rows("2008-12-24 21:30:00"));
 
   result =
@@ -249,7 +249,7 @@ public class DateTimeFunctionIT extends PPLIntegTestCase {
           "source=%s | eval f = DATETIME('2008-12-25 05:30:00+00:00', '+01:00') | fields f",
                TEST_INDEX_DATE));
   verifySchema(result,
-               schema("f", null, "timestamp"));
+        schema("f", null, "timestamp"));
   verifySome(result.getJSONArray("datarows"), rows("2008-12-25 06:30:00"));
 
   result =
@@ -257,7 +257,7 @@ public class DateTimeFunctionIT extends PPLIntegTestCase {
           "source=%s | eval f = DATETIME('2008-12-25 05:30:00-05:00', '+05:00') | fields f",
                TEST_INDEX_DATE));
   verifySchema(result,
-               schema("f", null, "timestamp"));
+        schema("f", null, "timestamp"));
   verifySome(result.getJSONArray("datarows"), rows("2008-12-25 15:30:00"));
 
   result =
@@ -265,7 +265,7 @@ public class DateTimeFunctionIT extends PPLIntegTestCase {
           "source=%s | eval f = DATETIME('2004-02-28 23:00:00-10:00', '+10:00') | fields f",
                TEST_INDEX_DATE));
   verifySchema(result,
-               schema("f", null, "timestamp"));
+        schema("f", null, "timestamp"));
   verifySome(result.getJSONArray("datarows"), rows("2004-02-29 19:00:00"));
 
   result =
@@ -273,7 +273,7 @@ public class DateTimeFunctionIT extends PPLIntegTestCase {
           "source=%s | eval f = DATETIME('2003-02-28 23:00:00-10:00', '+10:00') | fields f",
                TEST_INDEX_DATE));
   verifySchema(result,
-               schema("f", null, "timestamp"));
+        schema("f", null, "timestamp"));
   verifySome(result.getJSONArray("datarows"), rows("2003-03-01 19:00:00"));
 
   result =
@@ -281,7 +281,7 @@ public class DateTimeFunctionIT extends PPLIntegTestCase {
           "source=%s | eval f = DATETIME('2008-12-25 05:30:00+00:00', '+14:00') | fields f",
                TEST_INDEX_DATE));
   verifySchema(result,
-               schema("f", null, "timestamp"));
+        schema("f", null, "timestamp"));
   verifySome(result.getJSONArray("datarows"), rows("2008-12-25 19:30:00"));
 
   result =
@@ -289,7 +289,7 @@ public class DateTimeFunctionIT extends PPLIntegTestCase {
           "source=%s | eval f = DATETIME('2008-01-01 02:00:00+10:00', '-10:00') | fields f",
                TEST_INDEX_DATE));
   verifySchema(result,
-               schema("f", null, "timestamp"));
+        schema("f", null, "timestamp"));
   verifySome(result.getJSONArray("datarows"), rows("2007-12-31 06:00:00"));
 
   result =
@@ -297,7 +297,7 @@ public class DateTimeFunctionIT extends PPLIntegTestCase {
           "source=%s | eval f = DATETIME('2008-01-01 02:00:00+10:00') | fields f",
                TEST_INDEX_DATE));
   verifySchema(result,
-               schema("f", null, "timestamp"));
+        schema("f", null, "timestamp"));
   verifySome(result.getJSONArray("datarows"), rows("2008-01-01 02:00:00"));
 
   result =
@@ -305,7 +305,7 @@ public class DateTimeFunctionIT extends PPLIntegTestCase {
           "source=%s | eval f = DATETIME('2008-01-01 02:00:00') | fields f",
                TEST_INDEX_DATE));
   verifySchema(result,
-               schema("f", null, "timestamp"));
+        schema("f", null, "timestamp"));
   verifySome(result.getJSONArray("datarows"), rows("2008-01-01 02:00:00"));
 
   result =
@@ -313,7 +313,7 @@ public class DateTimeFunctionIT extends PPLIntegTestCase {
           "source=%s | eval f = DATETIME('2008-01-01 02:00:00+15:00', '-12:00') | fields f",
                TEST_INDEX_DATE));
   verifySchema(result,
-               schema("f", null, "timestamp"));
+        schema("f", null, "timestamp"));
   verifySome(result.getJSONArray("datarows"), rows(new Object[]{null}));
 
   result =
@@ -321,7 +321,7 @@ public class DateTimeFunctionIT extends PPLIntegTestCase {
           "source=%s | eval f = DATETIME('2008-01-01 02:00:00+10:00', '-14:00') | fields f",
                TEST_INDEX_DATE));
   verifySchema(result,
-               schema("f", null, "timestamp"));
+        schema("f", null, "timestamp"));
   verifySome(result.getJSONArray("datarows"), rows(new Object[]{null}));
 
   result =
@@ -329,10 +329,9 @@ public class DateTimeFunctionIT extends PPLIntegTestCase {
           "source=%s | eval f = DATETIME('2008-01-01 02:00:00', '-14:00') | fields f",
                TEST_INDEX_DATE));
   verifySchema(result,
-               schema("f", null, "timestamp"));
+        schema("f", null, "timestamp"));
   verifySome(result.getJSONArray("datarows"), rows(new Object[]{null}));
 }
-
 
   @Test
   public void testDateSub() throws IOException {
